Refactoring init/test script to be multiple files, and typescript.
