+++ bin/misc/clear-all.bat
