+++ bin/misc/merge.sh
