+++ package.json
@@ -9,11 +9,13 @@
   "name": "travetto-mono-repo",
   "license": "MIT",
   "devDependencies": {
-    "@types/faker": "^4.1.4",
+    "@types/faker": "^4.1.5",
     "@types/passport-facebook": "^2.1.8",
     "faker": "^4.1.0",
-    "lerna": "^3.12.0",
-    "passport-facebook": "^2.1.1"
+    "lerna": "^3.13.0",
+    "passport-facebook": "^2.1.1",
+    "js-yaml": "^3.12.1",
+    "@types/js-yaml": "^3.12.0"
   },
   "dependencies": {
     "typescript": "^3.3.1"
