+++ tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "compilerOptions": {
+    "module": "commonjs",
+    "target": "es2018",
+    "strict": true,
+    "strictPropertyInitialization": false,
+    "sourceMap": true,
+    "lib": [
+      "es2015.proxy",
+      "es2018"
+    ],
+    "preserveConstEnums": true,
+    "experimentalDecorators": true,
+    "importHelpers": true,
+    "rootDir": "bin/",
+    "outDir": "build/"
+  }
+}
